backend/requirements.txt
pyspark==3.5.1
duckdb==1.1.3
polars==1.12.0
pyarrow==17.0.0
numpy==1.26.4
scipy==1.13.1
pyyaml==6.0.2
h3==4.1.0
tqdm==4.66.5
click==8.1.7
